Clarify unfiltered_html capability in Multisite
In Multisite, regular Administrators of a single site do _not_ have `unfiltered_html`, only super-administrator accounts do.

Historically, WordPress referred to a single site as a "blog" and a group of sites in a multisite network as a "site", but this is no longer the contemporary usage. A single site is now referred to as a "site", and a group of sites is referred to as a "network".